Another attempt to fix compilation issue

+++ cocos/ui/UILayout.cpp
@@ -485,8 +485,8 @@ const Rect& Layout::getClippingRect()
 
         //Node can be flipped
         const auto worldPos = Vec2(std::min(worldPos1.x, worldPos2.x), std::min(worldPos1.y, worldPos2.y));
-        const auto scissorWidth = std::fabsf(worldPos2.x - worldPos1.x);
-        const auto scissorHeight = std::fabsf(worldPos2.y - worldPos1.y);
+        const auto scissorWidth = std::fabs(worldPos2.x - worldPos1.x);
+        const auto scissorHeight = std::fabs(worldPos2.y - worldPos1.y);
 
         Layout* parent = this;
